+++ src/aux-runtime/runtime/AuxLibrary.spec.ts
@@ -2520,6 +2520,28 @@ describe('AuxLibrary', () => {
                 expect(context.actions).toEqual([expected]);
             });
 
+            it('should include recordName in options', () => {
+                const promise: any = library.api.ai.chat('hello, world!', {
+                    recordName: 'myRecord',
+                });
+
+                const expected = aiChat(
+                    [
+                        {
+                            role: 'user',
+                            content: 'hello, world!',
+                        },
+                    ],
+                    {
+                        recordName: 'myRecord',
+                    },
+                    context.tasks.size
+                );
+
+                expect(promise[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
             it('should return the first chat choice', async () => {
                 let result: any;
                 const promise: any = library.api.ai.chat({
@@ -2705,6 +2727,31 @@ describe('AuxLibrary', () => {
                     {
                         preferredModel: 'gpt-3.5-turbo',
                         temperature: 0.5,
+                    },
+                    context.iterableTasks.size
+                );
+
+                expect(promise[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should include recordName in options', () => {
+                const promise: any = library.api.ai.stream.chat(
+                    'hello, world!',
+                    {
+                        recordName: 'myRecord',
+                    }
+                );
+
+                const expected = aiChatStream(
+                    [
+                        {
+                            role: 'user',
+                            content: 'hello, world!',
+                        },
+                    ],
+                    {
+                        recordName: 'myRecord',
                     },
                     context.iterableTasks.size
                 );
@@ -3483,6 +3530,28 @@ describe('AuxLibrary', () => {
                 expect(context.actions).toEqual([expected]);
             });
 
+            it('should include recordName in options', () => {
+                const promise: any = library.api.ai.generateSkybox(
+                    'cartoon clouds',
+                    undefined,
+                    {
+                        recordName: 'myRecord',
+                    }
+                );
+
+                const expected = aiGenerateSkybox(
+                    'cartoon clouds',
+                    undefined,
+                    {
+                        recordName: 'myRecord',
+                    },
+                    context.tasks.size
+                );
+
+                expect(promise[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
             it('should resolve with the address that was generated', async () => {
                 let result: string | null = null;
                 const promise: any =
@@ -3598,6 +3667,25 @@ describe('AuxLibrary', () => {
                     {
                         prompt: 'cartoon clouds',
                         negativePrompt: 'realistic',
+                    },
+                    undefined,
+                    context.tasks.size
+                );
+
+                expect(promise[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should include recordName in request objects', () => {
+                const promise: any = library.api.ai.generateImage({
+                    prompt: 'cartoon clouds',
+                    recordName: 'myRecord',
+                });
+
+                const expected = aiGenerateImage(
+                    {
+                        prompt: 'cartoon clouds',
+                        recordName: 'myRecord',
                     },
                     undefined,
                     context.tasks.size